Create BatchUnitMenuViewController. Move unit action handlers
