+++ Assets/Scripts/Units/Movement/ActionHandlers/UnitDestinationSelector.cs
@@ -10,7 +10,7 @@ using Units.Actions;
 using UnityEngine;
 using ILogger = Logging.ILogger;
 
-namespace Units.Movement.Handlers {
+namespace Units.Movement.ActionHandlers {
     /// <summary>
     /// This action handler shows all possible destinations for a unit to move to.
     /// Action is confirmed when the user taps or clicks on a valid destination.
