+++ Assets/Scripts/Units/Movement/UnitMovementInstaller.cs
@@ -1,4 +1,5 @@
 using Units.Actions;
+using Units.Movement.ActionHandlers;
 using Units.Movement.Handlers;
 using Zenject;
