+++ Assets/Scripts/Units/Selection/BatchUnitMenuViewController.cs
@@ -1,5 +1,9 @@
 namespace Units.Selection {
     internal class BatchUnitMenuViewController {
-        
+        public void Show(IUnit[] units) {
+        }
+
+        public void Hide() {
+        }
     }
 }
